home: refactor component
